move and rename new testing package
